Added ability to connect with already paired device.
In checkBluetoothState(), added a step to connection process, wherein, if Bluetooth is already on, no device is connected, but there are bonded (paired) devices, a dialog will prompt the user to connect with one of the already paired devices from a list displayed. The user still has the option to pair with a new device if the right one isn't present.